Split out the view models formatting cleanup
