+++ Visualizer/Models/DieViewModel.cs
@@ -1,0 +1,13 @@
+﻿using DataFramework.Models;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+
+namespace Visualizer.Models
+{
+	public class DieViewModel
+	{
+		public string DieType { get; set; }
+		public int Quantity { get; set; }
+	}
+}
